# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountDecisionHistoryResponse": {
        "additionalProperties": false,
        "description": "Paginated immutable Decision and Policy history for one account.",
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DecisionPolicyHistoryItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceResponse"
          }
        },
        "required": [
          "workspace",
          "account_id",
          "items",
          "limit",
          "offset"
        ],
        "title": "AccountDecisionHistoryResponse",
        "type": "object"
      },
...
      "AccountStateDetailResponse": {
        "additionalProperties": false,
        "description": "One snapshot with normalized explanations and full input traces.",
        "properties": {
          "evaluator_manifest": {
            "$ref": "#/components/schemas/StateEvaluatorManifestResponse"
          },
          "fit_criteria": {
            "items": {
              "$ref": "#/components/schemas/FitCriterionTraceResponse"
            },
            "title": "Fit Criteria",
            "type": "array"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/AccountStateReasonResponse"
            },
            "title": "Reasons",
            "type": "array"
          },
          "relationship_evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Relationship Evidence",
            "type": "array"
          },
          "signal_evaluations": {
            "items": {
              "$ref": "#/components/schemas/SignalEvaluationTraceResponse"
            },
            "title": "Signal Evaluations",
            "type": "array"
          },
          "snapshot": {
            "$ref": "#/components/schemas/AccountStateSnapshotResponse"
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceResponse"
          }
        },
        "required": [
          "workspace",
          "snapshot",
          "evaluator_manifest",
          "reasons",
          "fit_criteria",
          "relationship_evidence",
          "signal_evaluations"
        ],
        "title": "AccountStateDetailResponse",
        "type": "object"
      },
...
      "AccountStateReasonResponse": {
        "additionalProperties": false,
        "description": "One ordered, facet-scoped explanation.",
        "properties": {
          "facet": {
            "$ref": "#/components/schemas/AccountStateFacet"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "reason_code": {
            "$ref": "#/components/schemas/AccountStateReasonCode"
          }
        },
        "required": [
          "facet",
          "position",
          "reason_code"
        ],
        "title": "AccountStateReasonResponse",
        "type": "object"
      },
      "AccountStateSnapshotResponse": {
        "additionalProperties": false,
        "description": "Immutable descriptive account-state snapshot.",
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_sufficiency": {
            "$ref": "#/components/schemas/AccountEvidenceSufficiency"
          },
          "fit_context": {
            "$ref": "#/components/schemas/AccountFitContext"
          },
          "input_hash": {
            "title": "Input Hash",
            "type": "string"
          },
          "relationship_state": {
            "$ref": "#/components/schemas/AccountRelationshipState"
          },
          "state_as_of": {
            "format": "date-time",
            "title": "State As Of",
            "type": "string"
          },
          "state_engine_version": {
            "title": "State Engine Version",
            "type": "string"
          },
          "state_snapshot_id": {
            "format": "uuid",
            "title": "State Snapshot Id",
            "type": "string"
          },
          "strategy_version_id": {
            "format": "uuid",
            "title": "Strategy Version Id",
            "type": "string"
          },
          "timing_state": {
            "$ref": "#/components/schemas/AccountTimingState"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "state_snapshot_id",
          "workspace_id",
          "account_id",
          "strategy_version_id",
          "state_as_of",
          "computed_at",
          "input_hash",
          "state_engine_version",
          "fit_context",
          "timing_state",
          "relationship_state",
          "evidence_sufficiency",
          "created_at"
        ],
        "title": "AccountStateSnapshotResponse",
        "type": "object"
      },
...
      "DecisionPolicyDetailResponse": {
        "additionalProperties": false,
        "description": "Detailed current or historical M1C read model.",
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/DecisionTraceResponse"
          },
          "disposition": {
            "$ref": "#/components/schemas/ProposedDispositionResponse"
          },
          "policy": {
            "$ref": "#/components/schemas/PolicyTraceResponse"
          },
          "state_snapshot": {
            "$ref": "#/components/schemas/AccountStateSnapshotResponse"
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceResponse"
          }
        },
        "required": [
          "state_snapshot",
          "decision",
          "policy",
          "disposition",
          "workspace"
        ],
        "title": "DecisionPolicyDetailResponse",
        "type": "object"
      },
      "DecisionPolicyHistoryItemResponse": {
        "additionalProperties": false,
        "description": "One immutable Decision/Policy pair anchored to an exact state snapshot.",
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/DecisionTraceResponse"
          },
          "disposition": {
            "$ref": "#/components/schemas/ProposedDispositionResponse"
          },
          "policy": {
            "$ref": "#/components/schemas/PolicyTraceResponse"
          },
          "state_snapshot": {
            "$ref": "#/components/schemas/AccountStateSnapshotResponse"
          }
        },
        "required": [
          "state_snapshot",
          "decision",
          "policy",
          "disposition"
        ],
        "title": "DecisionPolicyHistoryItemResponse",
        "type": "object"
      },
...
      "DecisionTraceResponse": {
        "additionalProperties": false,
        "description": "Decision evaluation with its exact version and ordered reasons.",
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/DecisionDefinitionResponse"
          },
          "evaluation": {
            "$ref": "#/components/schemas/DecisionEvaluationResponse"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/DecisionEvaluationReasonResponse"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "definition",
          "evaluation",
          "reasons"
        ],
        "title": "DecisionTraceResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyResult": {
        "description": "Deterministic gate applied independently of Decision desirability.",
        "enum": [
          "ALLOW",
          "REQUIRE_REVIEW",
          "BLOCK"
        ],
        "title": "PolicyResult",
        "type": "string"
      },
...
      "PolicyTraceResponse": {
        "additionalProperties": false,
        "description": "Policy evaluation with its exact version and ordered reasons.",
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/PolicyDefinitionResponse"
          },
          "evaluation": {
            "$ref": "#/components/schemas/PolicyEvaluationResponse"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/PolicyEvaluationReasonResponse"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "definition",
          "evaluation",
          "reasons"
        ],
        "title": "PolicyTraceResponse",
        "type": "object"
      },
      "ProposedDispositionResponse": {
        "additionalProperties": false,
        "description": "Non-executing composition of separate immutable Decision and Policy results.",
        "properties": {
          "decision_evaluation_id": {
            "format": "uuid",
            "title": "Decision Evaluation Id",
            "type": "string"
          },
          "external_action_authorized": {
            "const": false,
            "default": false,
            "title": "External Action Authorized",
            "type": "boolean"
          },
          "lifecycle": {
            "const": "PROPOSED_ONLY",
            "default": "PROPOSED_ONLY",
            "title": "Lifecycle",
            "type": "string"
          },
          "policy_evaluation_id": {
            "format": "uuid",
            "title": "Policy Evaluation Id",
            "type": "string"
          },
          "policy_result": {
            "$ref": "#/components/schemas/PolicyResult"
          },
          "proposed_response": {
            "$ref": "#/components/schemas/DecisionResult",
            "description": "Proposed response posture. ENGAGE means engagement merits consideration only."
          },
          "proposed_response_label": {
            "title": "Proposed Response Label",
            "type": "string"
          },
          "state_snapshot_id": {
            "format": "uuid",
            "title": "State Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "state_snapshot_id",
          "decision_evaluation_id",
          "policy_evaluation_id",
          "proposed_response",
          "proposed_response_label",
          "policy_result"
        ],
        "title": "ProposedDispositionResponse",
        "type": "object"
      },
...
      "SignalEvaluationTraceResponse": {
        "additionalProperties": false,
        "description": "Evaluation result and its complete relational evidence trace.",
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/SignalDefinitionResponse"
          },
          "evaluation": {
            "$ref": "#/components/schemas/SignalEvaluationResponse"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          }
        },
        "required": [
          "evaluation",
          "definition",
          "evidence"
        ],
        "title": "SignalEvaluationTraceResponse",
        "type": "object"
      },
...
      "StateEvaluatorManifestResponse": {
        "additionalProperties": false,
        "description": "Exact facet evaluator manifest for a state-engine version.",
        "properties": {
          "evidence_sufficiency": {
            "$ref": "#/components/schemas/StateEvaluatorReferenceResponse"
          },
          "fit": {
            "$ref": "#/components/schemas/StateEvaluatorReferenceResponse"
          },
          "relationship": {
            "$ref": "#/components/schemas/StateEvaluatorReferenceResponse"
          },
          "timing": {
            "$ref": "#/components/schemas/StateEvaluatorReferenceResponse"
          }
        },
        "required": [
          "fit",
          "timing",
          "relationship",
          "evidence_sufficiency"
        ],
        "title": "StateEvaluatorManifestResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkspaceResponse": {
        "additionalProperties": false,
        "description": "Public workspace boundary for the synthetic demo.",
        "properties": {
          "demo_as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo As Of"
          },
          "demo_mode": {
            "title": "Demo Mode",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "slug",
          "name",
          "demo_mode",
          "demo_as_of"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Typed service boundary for evidence-backed GTM decisions.",
    "title": "GTM State & Signal Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/accounts/{account_id}/decision": {
      "get": {
        "description": "Return persisted Decision and Policy; ENGAGE means consideration, never execution.",
        "operationId": "get_current_account_decision_api_v1_accounts__account_id__decision_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionPolicyDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Account Decision",
        "tags": [
          "M1C Decision and Policy read model"
        ]
      }
    },
    "/api/v1/accounts/{account_id}/decision/history": {
      "get": {
        "description": "Return previously materialized immutable Decision and Policy pairs only.",
        "operationId": "list_account_decision_history_api_v1_accounts__account_id__decision_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "strategy_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy Version Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDecisionHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Account Decision History",
        "tags": [
          "M1C Decision and Policy read model"
        ]
      }
    },
...
    "/api/v1/accounts/{account_id}/state": {
      "get": {
        "description": "Return the latest materialized snapshot for the semantic state_as_of.",
        "operationId": "get_current_account_state_api_v1_accounts__account_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountStateDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Account State",
        "tags": [
          "M1B.2 account-state read model"
        ]
      }
    },
...
    "/api/v1/decision-evaluations/{decision_evaluation_id}": {
      "get": {
        "description": "Return one immutable Decision, its Policy, and exact snapshot reference.",
        "operationId": "get_decision_evaluation_api_v1_decision_evaluations__decision_evaluation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "decision_evaluation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Decision Evaluation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionPolicyDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Decision Evaluation",
        "tags": [
          "M1C Decision and Policy read model"
        ]
      }
    },
...
    "/api/v1/policy-evaluations/{policy_evaluation_id}": {
      "get": {
        "description": "Return one immutable Policy, its Decision, and exact snapshot reference.",
        "operationId": "get_policy_evaluation_api_v1_policy_evaluations__policy_evaluation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "policy_evaluation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Policy Evaluation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionPolicyDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Policy Evaluation",
        "tags": [
          "M1C Decision and Policy read model"
        ]
      }
    },
...
  }
}